AI powered fake news detector/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.0.1
torchvision==0.15.2
transformers==4.35.2
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.2
numpy==1.24.3
requests==2.31.0